+++ features/commands/calculateFishingPetsPrices.js
@@ -1,19 +1,50 @@
-import { BOLD, COMMON, DARK_GRAY, EPIC, GOLD, GREEN, LEGENDARY, MYTHIC, RARE, RESET, UNCOMMON } from "../../constants/formatting";
-import { getAuctionItemPrices } from "../../utils/auctionPrices";
+import { BOLD, DARK_GRAY, GOLD, GREEN, LEGENDARY, RESET } from "../../constants/formatting";
+import { getAuctionItemPrices, getPetRarityCode } from "../../utils/auctionPrices";
 import { toShortNumber } from "../../utils/common";
 import { isInSkyblock } from "../../utils/playerState";
 
+const MAX_XP = 25353230;
 const PETS_TO_CHECK = [
-    `${LEGENDARY}Blue Whale`,
-    `${LEGENDARY}Flying Fish`,
-    `${LEGENDARY}Baby Yeti`,
-    `${LEGENDARY}Penguin`,
-    `${LEGENDARY}Spinosaurus`,
-    `${LEGENDARY}Megalodon`,
-    `${LEGENDARY}Ammonite`,
-    `${LEGENDARY}Squid`,
-    `${LEGENDARY}Dolphin`,
-    `${LEGENDARY}Reindeer`,
+    {
+        petDisplayName: `${LEGENDARY}Blue Whale`,
+        xpGainMultiplier: 1
+    },
+    {
+        petDisplayName: `${LEGENDARY}Flying Fish`,
+        xpGainMultiplier: 1
+    },
+    {
+        petDisplayName: `${LEGENDARY}Baby Yeti`,
+        xpGainMultiplier: 1
+    },
+    {
+        petDisplayName: `${LEGENDARY}Penguin`,
+        xpGainMultiplier: 1
+    },
+    {
+        petDisplayName: `${LEGENDARY}Spinosaurus`,
+        xpGainMultiplier: 1
+    },
+    {
+        petDisplayName: `${LEGENDARY}Megalodon`,
+        xpGainMultiplier: 1
+    },
+    {
+        petDisplayName: `${LEGENDARY}Ammonite`,
+        xpGainMultiplier: 1
+    },
+    {
+        petDisplayName: `${LEGENDARY}Squid`,
+        xpGainMultiplier: 1
+    },
+    {
+        petDisplayName: `${LEGENDARY}Dolphin`,
+        xpGainMultiplier: 1
+    },
+    {
+        petDisplayName: `${LEGENDARY}Reindeer`,
+        xpGainMultiplier: 2 // 2x faster to level up
+    },
 ];
 
 export function calculateFishingPetPrices() {
@@ -22,20 +53,10 @@ export function calculateFishingPetPrices() {
             return;
         }
         
-        const prices = PETS_TO_CHECK.map(petDisplayName => {
-            const petName = petDisplayName.removeFormatting();
-            const rarityColorCode = petDisplayName.substring(0, 2);
-            const rarityCode = (function (rarityColorCode) {
-                switch (rarityColorCode) {
-                    case COMMON: return 0;
-                    case UNCOMMON: return 1;
-                    case RARE: return 2;
-                    case EPIC: return 3;
-                    case LEGENDARY: return 4;
-                    case MYTHIC: return 5;
-                    default: return 0;
-                }
-            })(rarityColorCode);
+        const prices = PETS_TO_CHECK.map(pet => {
+            const petName = pet.petDisplayName.removeFormatting();
+            const rarityColorCode = pet.petDisplayName.substring(0, 2);
+            const rarityCode = getPetRarityCode(rarityColorCode);
     
             const level1ItemId = petName.split(' ').join('_').toUpperCase() + ';' + rarityCode; // FLYING_FISH;4
             const level1AuctionPrices = getAuctionItemPrices(level1ItemId);
@@ -46,17 +67,18 @@ export function calculateFishingPetPrices() {
             const level100Price = level100AuctionPrices?.lbin || 0;
 
             return {
-                petDisplayName: petDisplayName,
+                petDisplayName: pet.petDisplayName,
                 level1Price: level1Price,
                 level100Price: level100Price,
+                coinsPerXp: (level100Price / MAX_XP) * pet.xpGainMultiplier,
                 diff: level1Price && level100Price ? level100Price - level1Price : 0
             };       
-        }).sort((a, b) => b.diff - a.diff);
+        }).sort((a, b) => b.coinsPerXp - a.coinsPerXp);
 
         let message = `${GREEN}${BOLD}Pets level up prices:\n`;
-        message += `${DARK_GRAY}Profits for leveling up the fishing pets from level 1 to level 100\n`;
+        message += `${DARK_GRAY}Profits for leveling up the fishing pets from level 1 to level 100.\n`;
         for (let petInfo of prices) {
-            message += ` - ${petInfo.petDisplayName}${RESET}: ${GREEN}+${toShortNumber(petInfo.diff) || 'N/A'}${RESET} (${GOLD}${toShortNumber(petInfo.level1Price) || 'N/A'}${RESET} -> ${GOLD}${toShortNumber(petInfo.level100Price) || 'N/A'}${RESET})\n`;
+            message += ` - ${petInfo.petDisplayName}${RESET}: ${GREEN}+${toShortNumber(petInfo.diff) || 'N/A'}${RESET} (${GOLD}${toShortNumber(petInfo.level1Price) || 'N/A'}${RESET} -> ${GOLD}${toShortNumber(petInfo.level100Price) || 'N/A'}${RESET}) | ${GOLD}${petInfo.coinsPerXp.toFixed(2)} ${RESET}coins/XP\n`;
         }
         ChatLib.chat(message);
     } catch (e) {